refactor: S3Service 리팩토링
