+++ src/main/java/org/ahpuh/surf/post/controller/PostController.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.ahpuh.surf.category.dto.CategorySimpleDto;
 import org.ahpuh.surf.common.response.CursorResult;
 import org.ahpuh.surf.common.s3.S3Service;
-import org.ahpuh.surf.common.s3.S3Service.FileStatus;
+import org.ahpuh.surf.common.s3.S3ServiceImpl.FileStatus;
 import org.ahpuh.surf.jwt.JwtAuthentication;
 import org.ahpuh.surf.post.dto.*;
 import org.ahpuh.surf.post.service.PostService;
