+++ src/main/java/org/ahpuh/surf/post/entity/Post.java
@@ -8,7 +8,7 @@ import lombok.experimental.SuperBuilder;
 import org.ahpuh.surf.category.entity.Category;
 import org.ahpuh.surf.common.entity.BaseEntity;
 import org.ahpuh.surf.common.exception.EntityExceptionHandler;
-import org.ahpuh.surf.common.s3.S3Service.FileStatus;
+import org.ahpuh.surf.common.s3.S3ServiceImpl.FileStatus;
 import org.ahpuh.surf.like.entity.Like;
 import org.ahpuh.surf.user.entity.User;
 import org.hibernate.annotations.Where;
